# bytes specialization, optimised from iterator implementation