--- a/newsfragments/4417.changed.md
+++ b/newsfragments/4417.changed.md
@@ -1,0 +1,2 @@
+`IntoPyObject` impls for `Vec<u8>`, `&[u8]`, `[u8; N]`, `Cow<[u8]>` and `SmallVec<[u8; N]>` now
+convert into `PyBytes` rather than `PyList`.