--- a/src/conversions/smallvec.rs
+++ b/src/conversions/smallvec.rs
@@ -70,6 +70,7 @@ where
     ///
     /// [`PyBytes`]: crate::types::PyBytes
     /// [`PyList`]: crate::types::PyList
+    #[inline]
     fn into_pyobject(self, py: Python<'py>) -> Result<Self::Output, Self::Error> {
         <A::Item>::iter_into_pyobject(self, py, crate::conversion::private::Token)
     }