--- a/src/conversions/std/slice.rs
+++ b/src/conversions/std/slice.rs
@@ -32,6 +32,7 @@ where
     ///
     /// [`PyBytes`]: crate::types::PyBytes
     /// [`PyList`]: crate::types::PyList
+    #[inline]
     fn into_pyobject(self, py: Python<'py>) -> Result<Self::Output, Self::Error> {
         <&T>::iter_into_pyobject(self, py, crate::conversion::private::Token)
     }
@@ -91,6 +92,7 @@ where
     type Output = Bound<'py, Self::Target>;
     type Error = PyErr;
 
+    #[inline]
     fn into_pyobject(self, py: Python<'py>) -> Result<Self::Output, Self::Error> {
         <&T>::iter_into_pyobject(self.iter(), py, crate::conversion::private::Token)
     }